+++ example_files/mappers/Baseline.rb
@@ -278,7 +278,10 @@ module URBANopt
           'Service',
           'Uncovered Parking',
           'Covered Parking',
-          'Mixed use'
+          'Mixed use',
+          'Single-Family',
+          'Multifamily (2 to 4 units)',
+          'Multifamily (5 or more units)'
         ]
       end