+++ spec/uo_cli_spec.rb
@@ -31,12 +31,18 @@
 RSpec.describe URBANopt::CLI do
   test_directory = File.join('spec', 'test_directory')
   test_directory_res = File.join('spec', 'test_directory_res')
+  test_directory_elec = File.join('spec', 'test_directory_elec')
   test_scenario = File.join(test_directory, 'two_building_scenario.csv')
   test_scenario_res = File.join(test_directory_res, 'two_building_res.csv')
   test_reopt_scenario = File.join(test_directory, 'REopt_scenario.csv')
+  test_scenario_elec = File.join(test_directory_elec, 'electrical_scenario.csv')
+  test_ev_scenario = File.join(test_directory, 'two_building_ev_scenario.csv')
   test_feature = File.join(test_directory, 'example_project.json')
   test_feature_res = File.join(test_directory_res, 'example_project_combined.json')
-  call_cli = "ruby #{File.join('lib', 'uo_cli.rb')}"
+  test_feature_elec = File.join(test_directory_elec, 'example_project_with_electric_network.json')
+  test_validate_bounds = File.join(test_directory_res, 'out_of_bounds_validation.yaml')
+  test_reopt_scenario_assumptions_file = File.join(test_directory, 'reopt', 'multiPV_assumptions.json')
+  call_cli = "bundle exec uo"
 
   # Ensure clean slate for testing
   # +dir_or_file+ string - path to a file or folder
@@ -90,11 +96,17 @@ RSpec.describe URBANopt::CLI do
     end
 
     it 'creates an example project directory for combined residential and commercial workflow' do
-      system("#{call_cli} create --project-folder #{test_directory} --combined")
-      expect(File.exist?(File.join(test_directory, 'mappers/residential'))).to be true
-      expect(File.exist?(File.join(test_directory, 'example_project_combined.json'))).to be true
-      expect(File.exist?(File.join(test_directory, 'measures'))).to be true
-      expect(File.exist?(File.join(test_directory, 'resources'))).to be true
+      delete_directory_or_file(test_directory_res)
+      system("#{call_cli} create --project-folder #{test_directory_res} --combined")
+      expect(File.exist?(File.join(test_directory_res, 'mappers/residential'))).to be true
+      expect(File.exist?(test_feature_res)).to be true
+      expect(File.exist?(File.join(test_directory_res, 'measures'))).to be true
+      expect(File.exist?(File.join(test_directory_res, 'resources'))).to be true
+    end
+
+    it 'creates an example project directory with electrical network properties' do
+      system("#{call_cli} create --project-folder #{test_directory_elec} --electric")
+      expect(File.exist?(test_feature_elec)).to be true
     end
 
     it 'creates an empty project directory' do
@@ -139,6 +151,7 @@ RSpec.describe URBANopt::CLI do
       expect(File.exist?(File.join(test_directory, 'baseline_scenario.csv'))).to be false
       system("#{call_cli} create --scenario-file #{test_feature}")
       expect(File.exist?(File.join(test_directory, 'baseline_scenario.csv'))).to be true
+      expect(File.exist?(File.join(test_directory, 'evcharging_scenario.csv'))).to be true
     end
 
     it 'creates a scenario file for a single feature from a feature file' do
@@ -160,74 +173,76 @@ RSpec.describe URBANopt::CLI do
   context 'Run and work with a small simulation' do
     before :all do
       delete_directory_or_file(test_directory)
+      delete_directory_or_file(test_directory_res)
+      delete_directory_or_file(test_directory_elec)
       system("#{call_cli} create --project-folder #{test_directory}")
+      system("#{call_cli} create --project-folder #{test_directory_res} --combined")
+      system("#{call_cli} create --project-folder #{test_directory_elec} --electric")
     end
 
     it 'runs a 2 building scenario using default geometry method' do
-      # Copy in a scenario file with only the first 2 buildings in it
+      # Use a ScenarioFile with only 2 buildings to reduce test time
       system("cp #{File.join('spec', 'spec_files', 'two_building_scenario.csv')} #{test_scenario}")
       system("#{call_cli} run --scenario #{test_scenario} --feature #{test_feature}")
-      expect(File.exist?(File.join(test_directory, 'run', 'two_building_scenario', '1', 'failed.job'))).to be false
-      expect(File.exist?(File.join(test_directory, 'run', 'two_building_scenario', '1', 'finished.job'))).to be true
+      expect(File.exist?(File.join(test_directory, 'run', 'two_building_scenario', '2', 'failed.job'))).to be false
+      expect(File.exist?(File.join(test_directory, 'run', 'two_building_scenario', '2', 'finished.job'))).to be true
       expect(File.exist?(File.join(test_directory, 'run', 'two_building_scenario', '3', 'finished.job'))).to be false
     end
 
     it 'runs a 2 building scenario with residential and commercial buildings' do
-      system("#{call_cli} create --project-folder #{test_directory_res} --combined")
       system("cp #{File.join('spec', 'spec_files', 'two_building_res.csv')} #{test_scenario_res}")
-
       system("#{call_cli} run --scenario #{test_scenario_res} --feature #{test_feature_res}")
-      expect(File.exist?(File.join(test_directory_res, 'run', 'two_building_res', '1', 'finished.job'))).to be true
+      expect(File.exist?(File.join(test_directory_res, 'run', 'two_building_res', '5', 'finished.job'))).to be true
       expect(File.exist?(File.join(test_directory_res, 'run', 'two_building_res', '16', 'finished.job'))).to be true
     end
 
     it 'runs a 2 building scenario using create bar geometry method' do
-      # Copy create bar mapper
+      # Copy create bar specific files
       system("cp #{File.join('example_files', 'mappers', 'CreateBar.rb')} #{File.join(test_directory, 'mappers', 'CreateBar.rb')}")
-      # Copy osw file
       system("cp #{File.join('example_files', 'mappers', 'createbar_workflow.osw')} #{File.join(test_directory, 'mappers', 'createbar_workflow.osw')}")
-      # Copy scenario file with 2 buildings in it
       system("cp #{File.join('spec', 'spec_files', 'two_building_create_bar.csv')} #{File.join(test_directory, 'two_building_create_bar.csv')}")
       system("#{call_cli} run --scenario #{File.join(test_directory, 'two_building_create_bar.csv')} --feature #{test_feature}")
-      expect(File.exist?(File.join(test_directory, 'run', 'two_building_create_bar', '1', 'finished.job'))).to be true
+      expect(File.exist?(File.join(test_directory, 'run', 'two_building_create_bar', '2', 'finished.job'))).to be true
     end
 
     it 'runs a 2 building scenario using floorspace geometry method' do
-      # Copy floorspace mapper
+      # Copy floorspace specific files
       system("cp #{File.join('example_files', 'mappers', 'Floorspace.rb')} #{File.join(test_directory, 'mappers', 'Floorspace.rb')}")
-      # Copy osw file
       system("cp #{File.join('example_files', 'mappers', 'floorspace_workflow.osw')} #{File.join(test_directory, 'mappers', 'floorspace_workflow.osw')}")
-      # Copy seed model and floospace.js file
       system("cp #{File.join('example_files', 'osm_building', '7_floorspace.json')} #{File.join(test_directory, 'osm_building', '7_floorspace.json')}")
       system("cp #{File.join('example_files', 'osm_building', '7_floorspace.osm')} #{File.join(test_directory, 'osm_building', '7_floorspace.osm')}")
-      # Copy feature file
       system("cp #{File.join('example_files', 'example_floorspace_project.json')} #{File.join(test_directory, 'example_floorspace_project.json')}")
-      # Copy scenario file with 2 buildings in it
       system("cp #{File.join('spec', 'spec_files', 'two_building_floorspace.csv')} #{File.join(test_directory, 'two_building_floorspace.csv')}")
       system("#{call_cli} run --scenario #{File.join(test_directory, 'two_building_floorspace.csv')} --feature #{File.join('../example_files/example_floorspace_project.json')}")
-      expect(File.exist?(File.join(test_directory, 'run', 'two_building_floorspace', '1', 'finished.job'))).to be true
+      expect(File.exist?(File.join(test_directory, 'run', 'two_building_floorspace', '5', 'finished.job'))).to be true
       expect(File.exist?(File.join(test_directory, 'run', 'two_building_floorspace', '7', 'finished.job'))).to be true
     end
 
+    it 'runs an ev-charging scenario' do
+      # copy ev-charging specific files
+      system("cp #{File.join('spec', 'spec_files', 'two_building_ev_scenario.csv')} #{test_ev_scenario}")
+      system("#{call_cli} run --scenario #{test_ev_scenario} --feature #{test_feature}")
+      expect(File.exist?(File.join(test_directory, 'run', 'two_building_ev_scenario', '5', 'finished.job'))).to be true
+      expect(File.exist?(File.join(test_directory, 'run', 'two_building_ev_scenario', '2', 'finished.job'))).to be true
+    end
+
+    it 'runs an electrical network scenario' do
+      system("cp #{File.join('spec', 'spec_files', 'electrical_scenario.csv')} #{test_scenario_elec}")
+      system("#{call_cli} run --scenario #{test_scenario_elec} --feature #{test_feature_elec}")
+      expect(File.exist?(File.join(test_directory_elec, 'run', 'electrical_scenario', '13', 'finished.job'))).to be true
+    end
+
     it 'runs a scenario when called with reopt' do
-      # Copy in a scenario file with only the first 2 buildings in it
       system("cp #{File.join('spec', 'spec_files', 'REopt_scenario.csv')} #{test_reopt_scenario}")
       # Copy in reopt folder
       system("cp -R #{File.join('spec', 'spec_files', 'reopt')} #{File.join(test_directory, 'reopt')}")
       system("#{call_cli} run --reopt --scenario #{test_reopt_scenario} --feature #{test_feature}")
-      expect(File.exist?(File.join(test_directory, 'run', 'reopt_scenario', '1', 'finished.job'))).to be true
+      expect(File.exist?(File.join(test_directory, 'run', 'reopt_scenario', '5', 'finished.job'))).to be true
       expect(File.exist?(File.join(test_directory, 'run', 'reopt_scenario', '2', 'finished.job'))).to be true
       expect(File.exist?(File.join(test_directory, 'run', 'reopt_scenario', '3', 'finished.job'))).to be false
     end
 
-    it 'checks for python from opendss command' do
-      # for now just check that it does the system check
-      expect { system("#{call_cli} opendss --scenario #{test_scenario} --feature #{test_feature}") }
-        .to output(a_string_including('Checking system.....'))
-        .to_stdout_from_any_process
-    end
-
-    it 'post-processor exits gracefully if given an invalid type' do
+    it 'post-processor closes gracefully if given an invalid type' do
       # Type is totally random
       expect { system("#{call_cli} process --foobar --scenario #{test_scenario} --feature #{test_feature}") }
         .to output(a_string_including("unknown argument '--foobar'"))
@@ -242,22 +257,26 @@ RSpec.describe URBANopt::CLI do
         .to_stderr_from_any_process
     end
 
-    it 'post-processor exits gracefully if not given a type' do
+    it 'post-processor closes gracefully if not given a type' do
       expect { system("#{call_cli} process --scenario #{test_scenario} --feature #{test_feature}") }
         .to output(a_string_including('No valid process type entered'))
         .to_stderr_from_any_process
     end
 
-    it 'post-processes a scenario' do
-      filename = File.join(test_directory, 'run', 'two_building_scenario', 'default_scenario_report.csv')
+    it 'default post-processes a scenario' do
+      test_scenario_report = File.join(test_directory, 'run', 'two_building_scenario', 'default_scenario_report.csv')
       system("#{call_cli} process --default --scenario #{test_scenario} --feature #{test_feature}")
-      filename = File.join(test_directory, 'run', 'two_building_scenario', 'default_scenario_report.csv')
-      expect(`wc -l < #{filename}`.to_i).to be > 2
+      expect(`wc -l < #{test_scenario_report}`.to_i).to be > 2
       expect(File.exist?(File.join(test_directory, 'run', 'two_building_scenario', 'process_status.json'))).to be true
     end
 
+    it 'successfully gets results from the opendss cli' do
+      system("#{call_cli} process --default --scenario #{test_scenario_elec} --feature #{test_feature_elec}")
+      system("#{call_cli} opendss --scenario #{test_scenario_elec} --feature #{test_feature_elec} --start-time '2017/01/15 01:00:00' --end-time '2017/01/22 01:00:00'")
+      expect(File.exist?(File.join(test_directory_elec, 'run', 'electrical_scenario', 'opendss', 'profiles', 'load_1.csv'))).to be true
+    end
+
     it 'saves post-process output as a database file' do
-      filename = File.join(test_directory, 'run', 'two_building_scenario', 'default_scenario_report.csv')
       db_filename = File.join(test_directory, 'run', 'two_building_scenario', 'default_scenario_report.db')
       system("#{call_cli} process --default --with-database --scenario #{test_scenario} --feature #{test_feature}")
       expect(`wc -l < #{db_filename}`.to_i).to be > 20
@@ -270,6 +289,14 @@ RSpec.describe URBANopt::CLI do
       expect(File.exist?(File.join(test_directory, 'run', 'reopt_scenario', 'process_status.json'))).to be true
     end
 
+    it 'reopt post-processes a scenario with specified scenario assumptions file' do
+      expect { system("#{call_cli} process --reopt-scenario -a #{test_reopt_scenario_assumptions_file} --scenario #{test_reopt_scenario} --feature #{test_feature}") }
+        .to output(a_string_including('multiPV_assumptions.json'))
+        .to_stdout_from_any_process
+      expect(File.exist?(File.join(test_directory, 'run', 'reopt_scenario', 'scenario_optimization.json'))).to be true
+      expect(File.exist?(File.join(test_directory, 'run', 'reopt_scenario', 'process_status.json'))).to be true
+    end
+
     it 'reopt post-processes each feature' do
       system("#{call_cli} process --reopt-feature --scenario #{test_reopt_scenario} --feature #{test_feature}")
       expect(File.exist?(File.join(test_directory, 'run', 'reopt_scenario', 'feature_optimization.csv'))).to be true
@@ -277,15 +304,16 @@ RSpec.describe URBANopt::CLI do
     end
 
     it 'opendss post-processes a scenario' do
-      expect(File.exist?(File.join(test_directory, 'run', 'two_building_scenario', 'opendss'))).to be false
-      system("cp -R #{File.join('spec', 'spec_files', 'opendss')} #{File.join(test_directory, 'run', 'two_building_scenario', 'opendss')}")
-      system("#{call_cli} process --opendss --scenario #{test_scenario} --feature #{test_feature}")
-      expect(File.exist?(File.join(test_directory, 'run', 'two_building_scenario', '1', 'feature_reports', 'default_feature_report_opendss.csv'))).to be true
-      expect(File.exist?(File.join(test_directory, 'run', 'two_building_scenario', 'process_status.json'))).to be true
+      expect(File.exist?(File.join(test_directory_elec, 'run', 'electrical_scenario', '2', 'feature_reports', 'default_feature_report_opendss.csv'))).to be false
+      system("#{call_cli} process --opendss --scenario #{test_scenario_elec} --feature #{test_feature_elec}")
+      expect(File.exist?(File.join(test_directory_elec, 'run', 'electrical_scenario', '2', 'feature_reports', 'default_feature_report_opendss.csv'))).to be true
+      expect(File.exist?(File.join(test_directory_elec, 'run', 'electrical_scenario', 'process_status.json'))).to be true
     end
 
     it 'creates scenario visualization for default post processor' do
+      # visualizing via the FeatureFile will throw error to stdout (but not crash) if a scenario that uses those features isn't processed first.
       system("#{call_cli} process --default --scenario #{test_scenario} --feature #{test_feature}")
+      system("#{call_cli} process --default --scenario #{test_ev_scenario} --feature #{test_feature}")
       system("#{call_cli} visualize --feature #{test_feature}")
       expect(File.exist?(File.join(test_directory, 'run', 'scenario_comparison.html'))).to be true
     end
@@ -313,11 +341,28 @@ RSpec.describe URBANopt::CLI do
       expect(File.exist?(File.join(test_directory, 'run', 'two_building_scenario', 'feature_comparison.html'))).to be true
     end
 
+    it 'validates eui' do
+      test_validation_file = File.join(test_directory_res, 'validation_schema.yaml')
+      expect { system("#{call_cli} validate --eui #{test_validation_file} --scenario #{test_scenario_res} --feature #{test_feature_res}") }
+        .to output(a_string_including('is within bounds set by'))
+        .to_stdout_from_any_process
+      system("cp #{File.join('spec', 'spec_files', 'out_of_bounds_validation.yaml')} #{test_validate_bounds}")
+      expect { system("#{call_cli} validate --eui #{test_validate_bounds} --scenario #{test_scenario_res} --feature #{test_feature_res}") }
+        .to output(a_string_including('kBtu/ft2/yr is greater than the validation maximum.'))
+        .to_stdout_from_any_process
+      expect { system("#{call_cli} validate --eui #{test_validate_bounds} --scenario #{test_scenario_res} --feature #{test_feature_res}") }
+        .to output(a_string_including('is less than the validation minimum.'))
+        .to_stdout_from_any_process
+      expect { system("#{call_cli} validate --eui #{test_validate_bounds} --scenario #{test_scenario_res} --feature #{test_feature_res} --units SI") }
+        .to output(a_string_including('kWh/m2/yr is less than the validation minimum.'))
+        .to_stdout_from_any_process
+    end
+
     it 'deletes a scenario' do
-      expect(File.exist?(File.join(test_directory, 'run', 'two_building_create_bar', '1', 'data_point_out.json'))).to be true
+      expect(File.exist?(File.join(test_directory, 'run', 'two_building_create_bar', '2', 'data_point_out.json'))).to be true
       bar_scenario = File.join(test_directory, "two_building_create_bar.csv")
       system("#{call_cli} delete --scenario #{bar_scenario}")
-      expect(File.exist?(File.join(test_directory, 'run', 'two_building_create_bar', '1', 'data_point_out.json'))).to be false
+      expect(File.exist?(File.join(test_directory, 'run', 'two_building_create_bar', '2', 'data_point_out.json'))).to be false
     end
   end
 end